requirements.txt
discord.py
motor
dnspython
openai
python-dotenv